% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{OptimFunction}
\alias{OptimFunction}
\title{OptimFunction}
\usage{
OptimFunction(param = c(0.6), d, s, doctorVisitingProb)
}
\arguments{
\item{param}{a}

\item{d}{a}

\item{s}{a}

\item{doctorVisitingProb}{a}
}
\description{
OptimFunction
}
