% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SetupStartInfected}
\alias{SetupStartInfected}
\title{SetupStartInfected}
\usage{
SetupStartInfected(d, s, startWeek, doctorVisitingProb,
  initialStartingMultiplier = 2)
}
\arguments{
\item{d}{a}

\item{s}{a}

\item{startWeek}{a}

\item{doctorVisitingProb}{a}

\item{initialStartingMultiplier}{a}
}
\description{
SetupStartInfected
}
